Fix generated model equality check implementation
The generated models include an `__eq__` implementation. This
implementation did no validation of the `other` argument and blindly
tried accessing its `__dict__` property. This property does not exist
for all values in Python, and as such the check would crash when
checking against these values.

This adds a simple sanity check at the beginning of each `__eq__`
method, verifying that the `other` argument is an instance of the model.
If not it is obviously not equal to the model instance running the
check.