Allow setting read only properties on models
When the API defines certain properties as read only, meaning any
updates to instances through the API won't accept values for those
properties, the generated models did not permit setting values for these
properties at all.

This made the generated models with properties like this broken, since
the `ApiClient` would not be able to populate new instances with data
from API responses.

A fix for this will be released in v2.2.2 of `swagger-codegen`. Until it
is released, this patches our models with read only properties in the
same way as it will be generated starting from v2.2.2.