Build based on default Swagger model template
We used to have our custom model template, but this broke with newer
Swagger Codegen in 0.3.0 and 0.3.1. Newer Swagger Codegen seems to have
fixed the issues we needed to solve with our custom template, so we can
now remove our model template completely. This build is then from the
default template.